+++ src/main/java/com/example/nutriplan_backend/repository/MacroClasificacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MacroClasificacionRepository extends JpaRepository<MacroClasificacion, Long> {
+public interface MacroClasificacionRepository extends JpaRepository<MacroClasificacion, Integer> {
 }
